LayoutTests/media/non-existent-video-playback-interrupted-expected.txt


RUN(internals.setMediaSessionRestrictions("videoaudio", "InterruptedPlaybackNotPermitted"))
RUN(video.src = "non-existent.mp4")
EXPECTED (video.paused == 'true') OK
RUN(internals.beginMediaSessionInterruption("System"))
RUN(video.play().then(didResolvePromise).catch(didRejectPromise))
RUN(internals.endMediaSessionInterruption("MayResumePlaying"))
Promise rejected. OK
EXPECTED (error.name == 'NotSupportedError') OK
END OF TEST
